+++ ios/COVIDSafePathsTests/ExposureManagerUnitTests.swift
@@ -2,7 +2,6 @@ import Foundation
 import UserNotifications
 import BackgroundTasks
 import ExposureNotification
-import RealmSwift
 import XCTest
 
 @testable import BT
@@ -738,325 +737,5 @@ class ExposureManagerTests: XCTestCase {
     }
     wait(for: [successExpetactionResolve, successExpectationReject], timeout: 0)
   }
-
-// MARK: == INTEGRATION TESTS ==
-  let indexTxt = """
-mn/1593432000-1593446400-00001.zip
-mn/1593432000-1593446400-00002.zip
-mn/1593432000-1593446400-00003.zip
-mn/1593432000-1593446400-00004.zip
-mn/1593432000-1593446400-00005.zip
-mn/1593432000-1593446400-00006.zip
-mn/1593432000-1593446400-00007.zip
-mn/1593432000-1593446400-00008.zip
-mn/1593432000-1593446400-00009.zip
-mn/1593432000-1593446400-00010.zip
-mn/1593432000-1593446400-00011.zip
-mn/1593432000-1593446400-00012.zip
-mn/1593432000-1593446400-00013.zip
-mn/1593432000-1593446400-00014.zip
-mn/1593432000-1593446400-00015.zip
-mn/1593432000-1593446400-00016.zip
-mn/1593432000-1593446400-00017.zip
-mn/1593432000-1593446400-00018.zip
-mn/1593432000-1593446400-00019.zip
-mn/1593432000-1593446400-00020.zip
-mn/1593432000-1593446400-00021.zip
-mn/1593432000-1593446400-00022.zip
-mn/1593432000-1593446400-00023.zip
-mn/1593446400-1593460800-00001.zip
-mn/1593446400-1593460800-00002.zip
-mn/1593446400-1593460800-00003.zip
-mn/1593446400-1593460800-00004.zip
-mn/1593446400-1593460800-00005.zip
-mn/1593446400-1593460800-00006.zip
-mn/1593446400-1593460800-00007.zip
-mn/1593446400-1593460800-00008.zip
-mn/1593446400-1593460800-00009.zip
-mn/1593446400-1593460800-00010.zip
-mn/1593446400-1593460800-00011.zip
-mn/1593446400-1593460800-00012.zip
-mn/1593446400-1593460800-00013.zip
-mn/1593446400-1593460800-00014.zip
-mn/1593446400-1593460800-00015.zip
-mn/1593446400-1593460800-00016.zip
-mn/1593446400-1593460800-00017.zip
-mn/1593446400-1593460800-00018.zip
-mn/1593446400-1593460800-00019.zip
-mn/1593446400-1593460800-00020.zip
-mn/1593446400-1593460800-00021.zip
-mn/1593446400-1593460800-00022.zip
-mn/1593446400-1593460800-00023.zip
-mn/1593460800-1593475200-00001.zip
-mn/1593460800-1593475200-00002.zip
-mn/1593460800-1593475200-00003.zip
-mn/1593460800-1593475200-00004.zip
-mn/1593460800-1593475200-00005.zip
-mn/1593460800-1593475200-00006.zip
-mn/1593460800-1593475200-00007.zip
-mn/1593460800-1593475200-00008.zip
-mn/1593460800-1593475200-00009.zip
-mn/1593460800-1593475200-00010.zip
-mn/1593460800-1593475200-00011.zip
-mn/1593460800-1593475200-00012.zip
-mn/1593460800-1593475200-00013.zip
-mn/1593460800-1593475200-00014.zip
-mn/1593460800-1593475200-00015.zip
-mn/1593460800-1593475200-00016.zip
-mn/1593460800-1593475200-00017.zip
-mn/1593460800-1593475200-00018.zip
-mn/1593460800-1593475200-00019.zip
-mn/1593460800-1593475200-00020.zip
-mn/1593460800-1593475200-00021.zip
-mn/1593460800-1593475200-00022.zip
-"""
-  
-  func testProcessesCorrectFilePathsDuringFirstRun() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = ""
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = Constants.dailyFileProcessingCapacity
-    let paths = exposureManager.urlPathsToProcess(indexTxt.gaenFilePaths)
-
-    XCTAssertEqual(paths.first!, "mn/1593432000-1593446400-00001.zip")
-    XCTAssertEqual(paths.last!, "mn/1593432000-1593446400-00015.zip")
-  }
-  
-  func testUrlPathsToProcessSecondPass() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    userState.urlOfMostRecentlyDetectedKeyFile = "mn/1593432000-1593446400-00015.zip"
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = "mn/1593432000-1593446400-00015.zip"
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = Constants.dailyFileProcessingCapacity
-    let paths = exposureManager.urlPathsToProcess(indexTxt.gaenFilePaths)
-
-    XCTAssertEqual(paths.first!, "mn/1593432000-1593446400-00016.zip")
-    XCTAssertEqual(paths.last!, "mn/1593446400-1593460800-00007.zip")
-  }
-  
-  func testUrlPathsToProcessAfterReadingAllButOneFile() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    userState.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00021.zip"
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00021.zip"
-    let paths = exposureManager.urlPathsToProcess(indexTxt.gaenFilePaths)
-
-    XCTAssertEqual(paths.count, 1)
-  }
-  
-  func testUrlPathsToProcessAfterReadingAllFiles() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    userState.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00022.zip"
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00022.zip"
-    let paths = exposureManager.urlPathsToProcess(indexTxt.gaenFilePaths)
-
-    XCTAssertEqual(paths.count, 0)
-  }
-  
-  func testUpdateRemainingFileCapacityFirstPass() {
-    let notificationCenter = NotificationCenter()
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: notificationCenter)
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock, notificationCenter: notificationCenter)
-
-    // Setup
-//    let userState = UserState()
-//    btSecureStorageMock.userStateHandler = {
-//      return userState
-//    }
-//    exposureManager.updateRemainingFileCapacity()
-//    let hoursSinceLastReset = Date.hourDifference(from: btSecureStorageMock.userState.dateLastPerformedFileCapacityReset!, to: Date())
-//
-//    XCTAssertEqual(hoursSinceLastReset, 0)
-//    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, Constants.dailyFileProcessingCapacity)
-  }
-  
-  func testUpdateRemainingFileCapacityUnder24Hours() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    userState.dateLastPerformedFileCapacityReset = Calendar.current.date(byAdding: .day, value: -2, to: Date())!
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.dateLastPerformedFileCapacityReset = Date()
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = 2
-    exposureManager.updateRemainingFileCapacity()
-    let hoursSinceLastReset = Date.hourDifference(from: btSecureStorageMock.userState.dateLastPerformedFileCapacityReset!,
-                                                  to: Date())
-
-    XCTAssertEqual(hoursSinceLastReset, 0)
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, 2)
-  }
-  
-  func testUpdateRemainingFileCapacityAfter24Hours() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let twoDaysAgo = Calendar.current.date(byAdding: .day, value: -2, to: Date())!
-    let userState = UserState()
-    userState.dateLastPerformedFileCapacityReset = Date()
-
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.dateLastPerformedFileCapacityReset = twoDaysAgo
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = 2
-    exposureManager.updateRemainingFileCapacity()
-    let hoursSinceLastReset = Date.hourDifference(from: btSecureStorageMock.userState.dateLastPerformedFileCapacityReset!,
-                                                  to: Date())
-
-    XCTAssertEqual(hoursSinceLastReset, 0)
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, Constants.dailyFileProcessingCapacity)
-  }
-  
-  func testSuccessfulDetection() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = Constants.dailyFileProcessingCapacity
-
-
-    exposureManager.finish(.success([]),
-                           processedFileCount: 4,
-                           lastProcessedUrlPath: "mn/1593460800-1593475200-00022.zip",
-                           progress: Progress()) { _ in }
-
-    // remainingDailyFileProcessingCapacity decreases, urlOfMostRecentlyDetectedKeyFile is stored
-    XCTAssertEqual(btSecureStorageMock.userState.urlOfMostRecentlyDetectedKeyFile, "mn/1593460800-1593475200-00022.zip")
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, 11)
-
-    // ----------------
-
-    exposureManager.finish(.success([]), processedFileCount: 8, lastProcessedUrlPath: "mn/1593460800-1593475200-00023.zip", progress: Progress()) { _ in }
-
-    // remainingDailyFileProcessingCapacity decreases, urlOfMostRecentlyDetectedKeyFile is stored
-    XCTAssertEqual(btSecureStorageMock.userState.urlOfMostRecentlyDetectedKeyFile, "mn/1593460800-1593475200-00023.zip")
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, 3)
-
-    // ----------------
-
-    // remainingDailyFileProcessingCapacity does not decrease, urlOfMostRecentlyDetectedKeyFile is not stored if empty string
-    exposureManager.finish(.success([]),
-                           processedFileCount: 0,
-                           lastProcessedUrlPath: .default,
-                           progress: Progress()) { _ in }
-
-    XCTAssertEqual(btSecureStorageMock.userState.urlOfMostRecentlyDetectedKeyFile, "mn/1593460800-1593475200-00023.zip")
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, 3)
-  }
-  
-  func testFailedDetection() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = Constants.dailyFileProcessingCapacity
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00022.zip"
-    exposureManager.finish(.failure(GenericError.unknown),
-                           processedFileCount: 4,
-                           lastProcessedUrlPath: "invalid",
-                           progress: Progress()) { _ in }
-
-    // remainingDailyFileProcessingCapacity does not decrease, urlOfMostRecentlyDetectedKeyFile is not stored
-    XCTAssertEqual(btSecureStorageMock.userState.urlOfMostRecentlyDetectedKeyFile, "mn/1593460800-1593475200-00022.zip")
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, Constants.dailyFileProcessingCapacity)
-  }
-  
-  func testCancelledDetection() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let progress = Progress()
-    progress.cancel()
-    btSecureStorageMock.remainingDailyFileProcessingCapacity = Constants.dailyFileProcessingCapacity
-    btSecureStorageMock.urlOfMostRecentlyDetectedKeyFile = "mn/1593460800-1593475200-00022.zip"
-    exposureManager.finish(.failure(GenericError.unknown),
-                           processedFileCount: 4,
-                           lastProcessedUrlPath: "invalid",
-                           progress: progress) { _ in }
-
-    // remainingDailyFileProcessingCapacity does not decrease, urlOfMostRecentlyDetectedKeyFile is not stored
-    XCTAssertEqual(btSecureStorageMock.userState.urlOfMostRecentlyDetectedKeyFile, "mn/1593460800-1593475200-00022.zip")
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity, Constants.dailyFileProcessingCapacity)
-  }
-
-  func testExposureStorage() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-    let exposureManager = ExposureManager(btSecureStorage: btSecureStorageMock)
-    // Setup
-    let userState = UserState()
-    btSecureStorageMock.userStateHandler = {
-      return userState
-    }
-    btSecureStorageMock.resetUserState() { _ in }
-
-    XCTAssertEqual(btSecureStorageMock.userState.exposures.count, 0)
-
-    let exposure = Exposure(id: UUID().uuidString,
-                            date: Date().posixRepresentation - Int(TimeInterval.random(in: 0...13)) * 24 * 60 * 60 * 1000,
-                            duration: TimeInterval(Int.random(in: 1...10) * 60 * 5 * 1000),
-                            totalRiskScore: .random(in: 1...8),
-                            transmissionRiskLevel: .random(in: 0...7))
-    exposureManager.finish(.success([exposure]),
-                           processedFileCount: 4,
-                           lastProcessedUrlPath: .default,
-                           progress: Progress()) { _ in }
-    XCTAssertEqual(btSecureStorageMock.userState.exposures.count, 1)
-  }
-
-  func testInitialFileCapacity() {
-    let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
-
-    // Setup
-    btSecureStorageMock.resetUserState() { _ in }
-    XCTAssertEqual(btSecureStorageMock.userState.remainingDailyFileProcessingCapacity,
-                   Constants.dailyFileProcessingCapacity)
-  }
-
-  func testFilterOldKeysForSubmission() {
-    // Keys older than 350 hours should not be submitted
-    
-    let oneMonthAgo: Date = Calendar.current.date(byAdding: .month, value: -1, to: Date())!
-    let oneWeekAgo: Date = Calendar.current.date(byAdding: .day, value: -7, to: Date())!
-    let today: Date = Date()
-
-    let keyA = ENTemporaryExposureKey()
-    keyA.rollingStartNumber = ENTemporaryExposureKey.rollingStartNumber(oneMonthAgo)
-
-    let keyB = ENTemporaryExposureKey()
-    keyB.rollingStartNumber = ENTemporaryExposureKey.rollingStartNumber(oneWeekAgo)
-
-    let keyC = ENTemporaryExposureKey()
-    keyC.rollingStartNumber = ENTemporaryExposureKey.rollingStartNumber(today)
-
-    let currentKeys = [keyA, keyB, keyC].current()
-
-    XCTAssertFalse(currentKeys.contains(keyA))
-    XCTAssertTrue(currentKeys.contains(keyB))
-    XCTAssertTrue(currentKeys.contains(keyC))
-  }
-
 }
-
